Refactor(cart): Enable decimal discounts and improve product sorting
This commit refactors the discount (`iskonto`) logic to support decimal values and enhances the product list's sorting algorithm.

### Core Changes:

-   **Decimal Discount Support (`iskonto`):**
    -   The `iskonto` field's data type has been changed from `int` to `double` across the entire application.
    -   This affects `CartProvider`, `CartItem`, refund models (`RefundItemModel`, `RefundListModel`), and database interactions.
    -   Input fields for discounts now accept decimal values, allowing for more precise pricing (e.g., `12.5%`).

-   **Enhanced Product Sorting (`cart_view.dart`):**
    -   Implemented a new `_sortProductsWithStock` method to provide a more intuitive product list order.
    -   Products are now sorted first alphabetically by name (A-Z) and then secondarily by stock quantity in descending order (items with more stock appear first).
    -   This ensures that products with the same name are grouped together, with available stock prioritized over suspended (out-of-stock) items.

### Detailed UI and Logic Updates:

-   **`cart_view.dart` & `cartsuggestion_view.dart` & `cart_view2.dart`**:
    -   Discount percentage calculations are updated to handle and store `double` values, preserving precision (e.g., `((original - new) / original * 100).toStringAsFixed(2)`).
    -   The keyboard type for discount input fields is now `TextInputType.numberWithOptions(decimal: true)` to improve user experience on both iOS and Android.
    -   Logic for applying, calculating, and clearing discounts now correctly uses `double` types (`0.0` instead of `0`).

-   **`cartsavedview.dart` (Saved Carts View):**
    -   Total calculation for saved carts now correctly incorporates the decimal discount value, ensuring accurate totals are displayed.
    -   The UI for individual items in a saved cart has been enhanced to clearly display discounts.
        -   If a discount is applied, the original price is shown with a line-through, next to the new, lower price.
        -   A red badge (e.g., "-15%") highlights the applied discount percentage.

-   **Data Models & Repositories (`refundsend_model.dart`, `refundlist_model.dart`, `refund_repository_impl.dart`, etc.)**:
    -   All models and data handling logic (JSON serialization, database mapping) have been updated to correctly process `iskonto` as a `double`.

+++ lib/features/cart/presentation/providers/cart_provider.dart
@@ -13,7 +13,7 @@ class CartItem {
   String birimTipi;
   int durum;
   String urunBarcode;
-  int iskonto;
+  double iskonto;
   String? imsrc;
 
   String adetFiyati; // yeni alan
@@ -121,13 +121,13 @@ class CartProvider extends ChangeNotifier {
 
   String get eskiFisNo => _eskiFisNo;
 
-  int getIskonto(String stokKodu, [String? birimTipi]) {
+  double getIskonto(String stokKodu, [String? birimTipi]) {
     if (birimTipi != null) {
       final cartKey = '${stokKodu}_$birimTipi';
-      return _items[cartKey]?.iskonto ?? 0;
+      return _items[cartKey]?.iskonto ?? 0.0;
     }
     // Backward compatibility: eski key ile dene
-    return _items[stokKodu]?.iskonto ?? 0;
+    return _items[stokKodu]?.iskonto ?? 0.0;
   }
 
   String getBirimTipi(String stokKodu, [String? birimTipi]) {
@@ -186,7 +186,7 @@ class CartProvider extends ChangeNotifier {
     required String urunBarcode,
     String? imsrc,
     int miktar = 1,
-    int iskonto = 0,
+    double iskonto = 0.0,
     String birimTipi = 'Box',
     int durum = 1,
     int vat = 18,
@@ -604,8 +604,8 @@ class CartProvider extends ChangeNotifier {
             : int.tryParse(item['miktar']?.toString() ?? '0') ?? 0,
         urunBarcode: item['urunBarcode']?.toString() ?? '',
         iskonto: (item['iskonto'] is num)
-            ? (item['iskonto'] as num).toInt()
-            : int.tryParse(item['iskonto']?.toString() ?? '0') ?? 0,
+            ? (item['iskonto'] as num).toDouble()
+            : double.tryParse(item['iskonto']?.toString() ?? '0') ?? 0.0,
         birimTipi: item['birimTipi']?.toString() ?? 'Box',
         durum: (item['durum'] is num)
             ? (item['durum'] as num).toInt()
